tao_triton/requirements-pip.txt
aiohttp==3.9.4
aiosignal==1.3.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
cuda-python==12.3.0
frozenlist==1.4.1
gevent==24.2.1
geventhttpclient==2.0.2
greenlet==3.0.3
grpcio==1.62.1
idna==3.7
imageio==2.34.0
joblib==1.4.0
kafka-python==2.0.2
lazy_loader==0.4
multidict==6.0.5
networkx==3.1
numpy==1.24.4
nvidia-pyindex==1.0.9
packaging==24.0
pillow==10.3.0
protobuf==3.20.3
python-rapidjson==1.16
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.4.0
tifffile==2023.7.10
tqdm==4.66.2
tritonclient==2.44.0
urllib3==2.2.1
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
